# optimize RNode.GetAnnotations and RNode.GetLabels