@ephesused 2022-10-13
After thinking this through, perhaps a better route would be for me to abandon this PR and come from the issue from another direction. My idea is to refactor rnode.go and fns.go so they use a single mapping node key/value traversal method:

1.  https://github.com/kubernetes-sigs/kustomize/blob/55a37de/kyaml/yaml/rnode.go#L423-L427
2.  https://github.com/kubernetes-sigs/kustomize/blob/55a37de/kyaml/yaml/rnode.go#L699-L701
3.  https://github.com/kubernetes-sigs/kustomize/blob/55a37de/kyaml/yaml/rnode.go#L712-L718
4.  https://github.com/kubernetes-sigs/kustomize/blob/55a37de/kyaml/yaml/rnode.go#L728-L733
5.  https://github.com/kubernetes-sigs/kustomize/blob/55a37de/kyaml/yaml/rnode.go#L1006-L1015
6. https://github.com/kubernetes-sigs/kustomize/blob/79a9154/kyaml/yaml/fns.go#L200-L227 (not sure that this one would fit the pattern)
7. https://github.com/kubernetes-sigs/kustomize/blob/79a9154/kyaml/yaml/fns.go#L405-L413
8. https://github.com/kubernetes-sigs/kustomize/blob/79a9154/kyaml/yaml/fns.go#L643-L649

That refactor would have value in and of itself, and it likely would set the stage for better consideration of how to manage the optimization that this PR intends to address.

What do you think?