@natasha41575 2022-12-21
@ephesused could we close this PR in favor of https://github.com/kubernetes-sigs/kustomize/pull/4944?