@ephesused 2022-12-21
> @ephesused could we close this PR in favor of #4944?

Definitely - I prefer that approach. Thanks for the prompt.